refactor([issue-2832]): split ensureSchemaImpl DDL into per-domain db/schema modules
Extract the ~1265-line inline DDL in ensureSchemaImpl() (server/lib/db.js)
into per-domain modules under server/lib/db/schema/, each exporting a plain
statement array; ensureSchemaImpl becomes a thin composer that runs
buildUpgradeDdl() then buildCatalogDdl(). Statement text and order are
byte-identical to the pre-split arrays (verified: 28 upgrade + 181 catalog
statements match exactly), so the composed schema is unchanged.

The catalog-DDL parity and catalogTypes text-scan tests now read the schema
module sources (the DDL literals moved out of db.js). Mirrors the
arcPlanner.js -> arcPlanner/* split (#1152).

Closes #2832
